requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
blis==0.7.10
catalogue==2.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
confection==0.1.1
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
cymem==2.0.7
decorator==5.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
executing==1.2.0
filelock==3.12.2
Flask==2.3.2
Flask-Cors==4.0.0
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
gdown==4.7.1
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jsonpickle==3.0.2
kiwisolver==1.4.5
langcodes==3.3.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
networkx==3.1
numpy==1.24.4
openai==0.27.8
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
plotly==5.16.0
preshed==3.0.8
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.12
Pygments==2.16.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyvis==0.3.2
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
smart-open==6.3.0
soupsieve==2.4.1
spacy==3.4.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
thinc==8.1.12
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.31.0
typer==0.7.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
validators==0.21.0
wasabi==0.10.1
wcwidth==0.2.6
weaviate-client==3.23.0
Werkzeug==2.3.7
yarl==1.9.2
zipp==3.16.2
